feat: adding client-side mdx rendering
We want to fetch the .mdx files from a remote source through the
useQuery hook, which means we cannot render it as easily as if were
importing the file directly from a local path.

The solution is to pass the raw text of the mdx file to the compile()
function with frontmatter plugins. Then, pass the result to the run()
function, which returns an MDXContent value that can be rendered.

Signed-off-by: David Dunham <dwdunham@proton.me>